+++ crates/objc2/src/protocol.rs
@@ -27,7 +27,7 @@ use crate::Message;
 /// # #[cfg(feature = "gnustep-1-7")]
 /// # unsafe { objc2::__gnustep_hack::get_class_to_force_linkage() };
 /// use objc2::ProtocolType;
-/// use objc2::foundation::NSObject;
+/// use objc2::runtime::NSObject;
 /// // Get a protocol object representing `NSObject`
 /// let protocol = NSObject::protocol();
 /// ```