+++ crates/objc2/src/runtime.rs
@@ -19,13 +19,17 @@ use std::os::raw::c_char;
 #[cfg(feature = "malloc")]
 use std::os::raw::c_uint;
 
+#[doc(hidden)]
+pub mod __nsstring;
 mod method_encoding_iter;
+mod nsobject;
 
 pub(crate) use self::method_encoding_iter::{EncodingParseError, MethodEncodingIter};
 use crate::encode::{Encode, EncodeArguments, EncodeConvert, Encoding, OptionEncode, RefEncode};
 use crate::ffi;
 use crate::verify::{verify_method_signature, Inner, VerificationError};
 
+pub use self::nsobject::NSObject;
 #[doc(inline)]
 pub use crate::encode::__bool::Bool;