+++ crates/test-ui/ui/extern_class_not_zst.rs
@@ -1,5 +1,5 @@
 use objc2::{extern_class, ClassType};
-use objc2::foundation::NSObject;
+use objc2::runtime::NSObject;
 
 extern_class!(
     pub struct NSNumber {