+++ crates/test-ui/ui/extern_methods_invalid_type.rs
@@ -1,5 +1,5 @@
 use objc2::{extern_class, extern_methods, ClassType};
-use objc2::foundation::NSObject;
+use objc2::runtime::NSObject;
 use objc2::rc::{Id, Owned};
 
 extern_class!(