+++ crates/test-ui/ui/msg_send_super_not_classtype.rs
@@ -1,7 +1,6 @@
 //! Invalid receiver to msg_send![super(obj), ...], missing ClassType impl.
 use objc2::msg_send;
-use objc2::foundation::NSObject;
-use objc2::runtime::Object;
+use objc2::runtime::{Object, NSObject};
 
 fn main() {
     let obj: &Object;