+++ crates/objc2/CHANGELOG.md
@@ -37,6 +37,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Added `ConformsTo` trait for marking that a type conforms to a specific
   protocol.
 * Added `Encode` impl for `Option<Sel>`.
+* Added `objc2::runtime::NSObject` - before, this was only available under
+  `objc2::foundation::NSObject`.
 
 ### Changed
 * Allow other types than `&Class` as the receiver in `msg_send_id!` methods